Updating colors to work with the MUI themes.
This is essentially a copy of #300 but updated to work with the
new layout. It also replaces all colors I could find to be sure
they refrence the themes and still look good on both light and
dark themes.

See #300 for what the screenshots look like as they are
an example. Also, I added a new theme example that will render
dark by default.