Fix issue where searchText did not show searches with page option
When the page option was set to something other than the first page, and the searchText option was used, searches did not show because the searches show only on the first page, and the page option defaulted to a different page. To preserve good UX, we should default back to the initial page when both options are set.