Fix invalid option deprecation checks
The `handleOptionDeprecation` function is being passed the original props instead of those with the default options merged in. This is causing false positive `console.error` output to be thrown.

Example:
Any table that doesn't include a `responsive` props option will see the `Invalid option value for responsive.` because the check isn't seeing the merged in default option.