Upgrade testing performance and simplify
This test sometimes times out when the test suite is run due to the way it is building new array data. It turns out that we don't need a new data set for this test, as we can use smaller rows per page number to get the same result on the existing data set in the test suite.