Prevent double sort from Enter key press
In the current implementation, using the enter key triggers the sort change twice. Once due to using enter key on a button, which by default (in Chrome at least) triggers that button's onClick, and the second time due to the onKeyUp function. Space bar is currently the work around, but for keyboard only users (blind/low vision) this is broken behavior.